Removes isFraction duplication
